feat: add delete method
